+++ tests/test_pytesseract.py
@@ -16,6 +16,10 @@ def test_pytesseract():
     # Tokens.
     assert doc[0].text == "Test"
     assert doc[-4].text == "data"
+    # Doc attributes.
+    assert doc._.page_range == (1, 4)
+    assert doc._.first_page == 1
+    assert doc._.last_page == 4
 
 
 def test_pytesseract_with_params():
@@ -30,3 +34,8 @@ def test_pytesseract_with_params():
     # Tokens.
     assert doc[0].text == "Test"
     assert doc[-4].text == "data"
+    # Doc attributes.
+    assert doc._.page_range == (1, 4)
+    assert doc._.first_page == 1
+    assert doc._.last_page == 4
+